backend/pyproject.toml
[project]
name = "slopahh-muzik-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
    "gunicorn>=21.2.0",
    "google-generativeai==0.3.2",
    "pydantic==2.5.3",
    "pytest==8.0.0",
    "python-dotenv==1.0.0",
    "requests>=2.32.5",
    "sqlalchemy==2.0.25",
]
